+++ src/freenet/clients/http/DiagnosticToadlet.java
@@ -20,15 +20,7 @@ import freenet.config.SubConfig;
 import freenet.io.xfer.BlockReceiver;
 import freenet.io.xfer.BlockTransmitter;
 import freenet.l10n.BaseL10n;
-import freenet.node.Node;
-import freenet.node.NodeClientCore;
-import freenet.node.NodeStarter;
-import freenet.node.NodeStats;
-import freenet.node.OpennetManager;
-import freenet.node.PeerManager;
-import freenet.node.PeerNodeStatus;
-import freenet.node.RequestTracker;
-import freenet.node.Version;
+import freenet.node.*;
 import freenet.node.stats.DataStoreInstanceType;
 import freenet.node.stats.DataStoreStats;
 import freenet.node.stats.StatsNotAvailableException;
@@ -64,7 +56,7 @@ public class DiagnosticToadlet extends Toadlet {
 		this.core = core;
 		this.fcp = fcp;
 		stats = node.nodeStats;
-		peers = node.peers;
+		peers = node.getPeerManager();
 		/* copied from NodeL10n constructor. */
 		baseL10n = new BaseL10n("freenet/l10n/", "freenet.l10n.${lang}.properties", new File(".").getPath()+File.separator+"freenet.l10n.${lang}.override.properties", BaseL10n.LANGUAGE.ENGLISH);
 	}
