+++ src/freenet/io/comm/IncomingPacketFilterImpl.java
@@ -58,7 +58,7 @@ public class IncomingPacketFilterImpl implements IncomingPacketFilter {
 	public DECODED process(byte[] buf, int offset, int length, Peer peer, long now) {
 		if(logMINOR) Logger.minor(this, "Packet length "+length+" from "+peer);
 		node.random.acceptTimerEntropy(fnpTimingSource, 0.25);
-		PeerNode opn = node.peers.getByPeer(peer, mangler);
+		PeerNode opn = node.getPeerManager().getByPeer(peer, mangler);
 
 		if(opn != null) {
 			if(opn.handleReceivedPacket(buf, offset, length, now, peer)) {
