+++ src/freenet/node/AnnounceSender.java
@@ -143,7 +143,7 @@ public class AnnounceSender implements PrioRunnable, ByteCounter {
 
 			if(onlyNode == null) {
 				// Route it
-				next = node.peers.closerPeer(source, nodesRoutedTo, target, true, node.isAdvancedModeEnabled(), -1,
+				next = node.getPeerManager().closerPeer(source, nodesRoutedTo, target, true, node.isAdvancedModeEnabled(), -1,
 				        null, null, htl, 0, source == null, false, false);
 			} else {
 				next = onlyNode;
