+++ src/freenet/node/CHKInsertSender.java
@@ -496,7 +496,7 @@ public final class CHKInsertSender extends BaseSender implements PrioRunnable, A
             
             // Route it
             // Can backtrack, so only route to nodes closer than we are to target.
-            next = node.peers.closerPeer(forkedRequestTag == null ? source : null, nodesRoutedTo, target, true, node.isAdvancedModeEnabled(), -1, null,
+            next = node.getPeerManager().closerPeer(forkedRequestTag == null ? source : null, nodesRoutedTo, target, true, node.isAdvancedModeEnabled(), -1, null,
 			        null, htl, ignoreLowBackoff ? Node.LOW_BACKOFF : 0, source == null, realTimeFlag, newLoadManagement);
 			
             if(next == null) {
