+++ src/freenet/node/DNSRequester.java
@@ -54,7 +54,7 @@ public class DNSRequester implements Runnable {
     }
 
     private void realRun() {
-        PeerNode[] nodes = node.peers.myPeers();
+        PeerNode[] nodes = ((ProtectedPeerManager) node.getPeerManager()).myPeers();
         long now = System.currentTimeMillis();
         if((now - lastLogTime) > 1000) {
         	if(logMINOR)
