+++ src/freenet/node/FNPPacketMangler.java
@@ -1323,7 +1323,7 @@ public class FNPPacketMangler implements OutgoingPacketMangler {
 					c, Ke, Ka, authenticator, hisRef, pn, replyTo, unknownInitiator, setupType, newTrackerID, newTrackerID == trackerID);
 
 			if(dontWant) {
-				node.peers.disconnectAndRemove(pn, true, true, true); // Let it connect then tell it to remove it.
+				node.getPeerManager().disconnectAndRemove(pn, true, true, true); // Let it connect then tell it to remove it.
 			} else {
 				pn.maybeSendInitialMessages();
 			}
@@ -1375,7 +1375,7 @@ public class FNPPacketMangler implements OutgoingPacketMangler {
 				Logger.normal(this, "Already connected to seednode");
 				return pn;
 			}
-			node.peers.addPeer(seed);
+			node.getPeerManager().addPeer(seed);
 			return seed;
 		} else {
 			Logger.error(this, "Unknown setup type");
@@ -1557,7 +1557,7 @@ public class FNPPacketMangler implements OutgoingPacketMangler {
 				pn.theirInitialMsgID);
 		if(newTrackerID >= 0) {
 			if(dontWant) {
-				node.peers.disconnectAndRemove(pn, true, true, true);
+				node.getPeerManager().disconnectAndRemove(pn, true, true, true);
 			} else {
 				pn.maybeSendInitialMessages();
 			}
