+++ src/freenet/node/NodeCrypto.java
@@ -306,7 +306,7 @@ public class NodeCrypto {
 		if(!(forARK || forSetup || forAnonInitiator)) {
 		    // We *do* need the location on noderefs exchanged via path folding and announcement.
 		    // This is necessary so we can take the location into account in OpennetManager.wantPeer().
-		    fs.put("location", node.lm.getLocation());
+		    fs.put("location", node.getLocationManager().getLocation());
 		}
 		fs.putSingle("version", Version.getVersionString()); // Keep, vital that peer know our version. For example, some types may be sent in different formats to different node versions (e.g. Peer).
 		if(!forAnonInitiator)
@@ -463,11 +463,11 @@ public class NodeCrypto {
 	}
 
 	public PeerNode[] getPeerNodes() {
-		if(node.peers == null) return null;
+		if(node.getPeerManager() == null) return null;
 		if(isOpennet)
-			return node.peers.getOpennetAndSeedServerPeers();
+			return node.getPeerManager().getOpennetAndSeedServerPeers();
 		else
-			return node.peers.getDarknetPeers();
+			return node.getPeerManager().getDarknetPeers();
 	}
 
 	public boolean allowConnection(PeerNode pn, FreenetInetAddress addr) {
@@ -475,7 +475,7 @@ public class NodeCrypto {
     		// Disallow multiple connections to the same address
 			// TODO: this is inadequate for IPv6, should be replaced by
 			// check for "same /64 subnet" [configurable] instead of exact match
-    		if(node.peers.anyConnectedPeerHasAddress(addr, pn) && !detector.includes(addr)
+    		if(node.getPeerManager().anyConnectedPeerHasAddress(addr, pn) && !detector.includes(addr)
     				&& addr.isRealInternetAddress(false, false, false)) {
     			Logger.normal(this, "Not sending handshake packets to "+addr+" for "+pn+" : Same IP address as another node");
     			return false;
@@ -493,7 +493,7 @@ public class NodeCrypto {
 			FreenetInetAddress address) {
 		if(detector.includes(address)) return;
 		if(!address.isRealInternetAddress(false, false, false)) return;
-		ArrayList<PeerNode> possibleMatches = node.peers.getAllConnectedByAddress(address, true);
+		ArrayList<PeerNode> possibleMatches = node.getPeerManager().getAllConnectedByAddress(address, true);
 		if(possibleMatches == null) return;
 		for(PeerNode pn : possibleMatches) {
 			if(pn == peerNode) continue;
@@ -510,7 +510,7 @@ public class NodeCrypto {
 					Logger.error(this, "Dropping peer "+pn+" because don't want connection due to others on the same IP address!");
 					System.out.println("Disconnecting permanently from your friend \""+((DarknetPeerNode)pn).getName()+"\" because your friend \""+((DarknetPeerNode)peerNode).getName()+"\" is using the same IP address "+address+"!");
 				}
-				node.peers.disconnectAndRemove(pn, true, true, pn.isOpennet());
+				node.getPeerManager().disconnectAndRemove(pn, true, true, pn.isOpennet());
 			}
 		}
 	}
@@ -524,7 +524,7 @@ public class NodeCrypto {
 
 	public PeerNode[] getAnonSetupPeerNodes() {
 		ArrayList<PeerNode> v = new ArrayList<PeerNode>();
-		for(PeerNode pn: node.peers.myPeers()) {
+		for(PeerNode pn: ((ProtectedPeerManager) node.getPeerManager()).myPeers()) {
 			if(pn.handshakeUnknownInitiator() && pn.getOutgoingMangler() == packetMangler)
 				v.add(pn);
 		}
