+++ src/freenet/node/PacketSender.java
@@ -91,7 +91,7 @@ public class PacketSender implements Runnable {
 					if (logMINOR)
 						Logger.minor(PacketSender.class,
 								"Starting shedulePeriodicJob() at " + now);
-					PeerManager pm = node.peers;
+					PeerManager pm = node.getPeerManager();
 					pm.maybeLogPeerNodeStatusSummary(now);
 					pm.maybeUpdateOldestNeverConnectedDarknetPeerAge(now);
 					stats.maybeUpdatePeerManagerUserAlertStats(now);
@@ -147,10 +147,10 @@ public class PacketSender implements Runnable {
 	 */
 	private void realRun() {
 		long now = System.currentTimeMillis();
-                PeerManager pm;
+		ProtectedPeerManager pm;
 		PeerNode[] nodes;
 
-        pm = node.peers;
+        pm = (ProtectedPeerManager) node.getPeerManager();
         nodes = pm.myPeers();
 
 		long nextActionTime = Long.MAX_VALUE;
@@ -205,7 +205,7 @@ public class PacketSender implements Runnable {
 			pn.maybeOnConnect();
 			if(pn.shouldDisconnectAndRemoveNow() && !pn.isDisconnecting()) {
 				// Might as well do it properly.
-				node.peers.disconnectAndRemove(pn, true, true, false);
+				node.getPeerManager().disconnectAndRemove(pn, true, true, false);
 			}
 
 			if(pn.isConnected()) {
@@ -227,7 +227,7 @@ public class PacketSender implements Runnable {
 					// Do it properly.
 					// There appears to be connectivity from them to us but not from us to them.
 					// So it is helpful for them to know that we are disconnecting.
-					node.peers.disconnect(pn, true, true, false, true, false, SECONDS.toMillis(5));
+					node.getPeerManager().disconnect(pn, true, true, false, true, false, SECONDS.toMillis(5));
 					continue;
 				} else if(pn.isRoutable() && pn.noLongerRoutable()) {
 					/*
