+++ src/freenet/node/ProtectedLocationManager.java
@@ -1,0 +1,7 @@
+package freenet.node;
+
+interface ProtectedLocationManager extends LocationManager {
+
+    void announceLocChange();
+
+}
