+++ src/freenet/node/RequestStarter.java
@@ -111,7 +111,7 @@ public class RequestStarter implements Runnable, RandomGrabArrayItemExclusionLis
 		while(true) {
 			// Allow 5 minutes before we start killing requests due to not connecting.
 			OpennetManager om;
-			if(core.node.peers.countConnectedPeers() < 3 && (om = core.node.getOpennet()) != null &&
+			if(core.node.getPeerManager().countConnectedPeers() < 3 && (om = core.node.getOpennet()) != null &&
 					System.currentTimeMillis() - om.getCreationTime() < MINUTES.toMillis(5)) {
 				try {
 					synchronized(this) {
