+++ src/freenet/node/SSKInsertSender.java
@@ -193,7 +193,7 @@ public class SSKInsertSender extends BaseSender implements PrioRunnable, AnyInse
             }
             
             // Route it
-            next = node.peers.closerPeer(forkedRequestTag == null ? source : null, nodesRoutedTo, target, true, node.isAdvancedModeEnabled(), -1, null,
+            next = node.getPeerManager().closerPeer(forkedRequestTag == null ? source : null, nodesRoutedTo, target, true, node.isAdvancedModeEnabled(), -1, null,
 			        null, htl, ignoreLowBackoff ? Node.LOW_BACKOFF : 0, source == null, realTimeFlag, newLoadManagement);
             
             if(next == null) {
