+++ src/freenet/node/SeedClientPeerNode.java
@@ -93,7 +93,7 @@ public class SeedClientPeerNode extends PeerNode {
 	@Override
 	public boolean disconnected(boolean dumpMessageQueue, boolean dumpTrackers) {
 		boolean ret = super.disconnected(true, true);
-		node.peers.disconnectAndRemove(this, false, false, false);
+		node.getPeerManager().disconnectAndRemove(this, false, false, false);
 		return ret;
 	}
