+++ src/freenet/node/ThrottleWindowManager.java
@@ -43,7 +43,7 @@ public class ThrottleWindowManager {
 		if (_simulatedWindowSize < 1.0) {
 			_simulatedWindowSize = 1.0F;
 		}
-		return _simulatedWindowSize * Math.max(1, node.peers.countNonBackedOffPeers(realTime));
+		return _simulatedWindowSize * Math.max(1, node.getPeerManager().countNonBackedOffPeers(realTime));
 	}
 
 	public synchronized void rejectedOverload() {
