+++ src/freenet/node/updater/NodeUpdateManager.java
@@ -29,7 +29,6 @@ import freenet.io.comm.Message;
 import freenet.io.comm.NotConnectedException;
 import freenet.keys.FreenetURI;
 import freenet.l10n.NodeL10n;
-import freenet.node.Announcer;
 import freenet.node.Node;
 import freenet.node.NodeInitException;
 import freenet.node.NodeFile;
@@ -574,7 +573,7 @@ public class NodeUpdateManager {
 			broadcastUOMAnnouncesOld = true;
 			msg = getOldUOMAnnouncement();
 		}
-		node.peers.localBroadcast(msg, true, true, ctr, 0, TRANSITION_VERSION-1);
+		node.getPeerManager().localBroadcast(msg, true, true, ctr, 0, TRANSITION_VERSION-1);
 	}
 
 	void broadcastUOMAnnouncesNew() {
@@ -588,7 +587,7 @@ public class NodeUpdateManager {
 			msg = getNewUOMAnnouncement(size);
 		}
 		if(logMINOR) Logger.minor(this, "Broadcasting UOM announcements (new)");
-		node.peers.localBroadcast(msg, true, true, ctr, TRANSITION_VERSION, Integer.MAX_VALUE);
+		node.getPeerManager().localBroadcast(msg, true, true, ctr, TRANSITION_VERSION, Integer.MAX_VALUE);
 	}
 
 	/** Return the length of the data fetched for the current version, or -1. */
@@ -1888,7 +1887,7 @@ public class NodeUpdateManager {
 			// Normally this means we won't send UOM.
 			// However, if something breaks severely, we need an escape route.
 			if (node.getUptime() > MINUTES.toMillis(5)
-					&& node.peers.countCompatibleRealPeers() == 0)
+					&& node.getPeerManager().countCompatibleRealPeers() == 0)
 				return false;
 			return true;
 		}
